Enhance financial tracking with BigDecimal precision, Glance widget, and background workers
- Migrate monetary fields to `BigDecimal` across entities, DAOs, repositories, and UI components with Room TypeConverters.
- Add month and year scoping to `Budget` entity and update `BudgetViewModel` to track monthly budget progress and status summaries.
- Add `isReviewed` transaction status, unreviewed transaction flow, and dedicated `TransactionReviewScreen`.
- Implement `ExpenditureWidget` using Jetpack Glance to display total balance and monthly net flow on the home screen.
- Set up WorkManager workers (`BudgetWorker`, `DailyReflectionWorker`) and `NotificationHelper` for budget alerts, daily reflections, and transaction review notifications.
- Enhance UI screens with custom charts, weekly transaction grouping, inline calculator text field (`CalculatorTextField`), and hierarchical category emoji picker (`CategoryTabSelector`).
- Add transaction CSV export functionality in `AnalyticsViewModel` via Android `FileProvider`.
- Improve Indian bank SMS parsing logic and regex patterns with extended unit test coverage.
- Refine app theme palette, typography, and adaptive launcher icons.

+++ app/src/main/java/com/example/myexpenditureapp/domain/model/BudgetWithProgress.kt
@@ -2,10 +2,11 @@ package com.example.myexpenditureapp.domain.model
 
 import com.example.myexpenditureapp.data.entity.Budget
 import com.example.myexpenditureapp.data.entity.Category
+import java.math.BigDecimal
 
 data class BudgetWithProgress(
     val budget: Budget,
     val category: Category?,
-    val currentSpending: Double,
+    val currentSpending: BigDecimal,
     val progress: Float // 0.0 to 1.0 or more
 )
